Rename TestClass and move to the controller package
Also added the @Controller annotation to facilitate the coming
autowired test.
